refactor(tui): move Textual UI into package
