+++ src/ursa/cli/tui/__init__.py
@@ -1,0 +1,1 @@
+"""Textual user interface for the URSA command-line application."""
