+++ src/ursa/cli/tui/agent_info.py
