+++ src/ursa/cli/tui/app.tcss
