+++ src/ursa/cli/tui/event_cards/base.py
