+++ src/ursa/cli/tui/event_cards/commands.py
@@ -11,8 +11,8 @@ from textual.app import ComposeResult
 from textual.containers import Horizontal
 from textual.widgets import Static
 
-from ursa.cli.event_cards.base import EventCard
-from ursa.cli.widgets import ActivityIndicator
+from ursa.cli.tui.event_cards.base import EventCard
+from ursa.cli.tui.widgets import ActivityIndicator
 
 
 class CommandSafetyIndicator(ActivityIndicator):
