+++ src/ursa/cli/tui/tips.py
